Separated plugin API from plugin service.
